% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{call_geno}
\alias{call_geno}
\title{Genotype with kmeans using three clusters}
\usage{
call_geno(theta, intensity, clusterm = c(0.2, 0.5, 0.8), delthresh = NULL)
}
\arguments{
\item{theta}{Vector with theta values.}

\item{intensity}{Vector with intensity values.}

\item{clusterm}{Vector containing the cluster means.}

\item{delthresh}{Threshold for intensity based deletion calling.}
}
\value{
List containing genotypes, b-allele frequencies and R ratios.
}
\description{
Genotype with kmeans using three clusters
}
\keyword{internal}

