% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filt_samp}
\alias{filt_samp}
\title{Filter Samples}
\usage{
filt_samp(dat, samples)
}
\arguments{
\item{dat}{raw_data object.}

\item{samples}{Character vector with sample names, prefixes of samples to
filter or sample indices.
If prefixes fit many samples, indices are highly recommended.}
}
\value{
raw_data object
}
\description{
Failed or negative control samples (e.g. H2O) should be filtered out before the normalization.
}
\examples{
if(require(brassicaData)){
data("raw_napus", package = "brassicaData", envir = environment())
raw_napus <- filt_samp(raw_napus, check_raw(raw = raw_napus, plot = FALSE, thresh = 28000))
}
}

