% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_cnv}
\alias{filter_cnv}
\title{Filter Copy number variations}
\usage{
filter_cnv(dat, thresh = 5)
}
\arguments{
\item{dat}{norm_data object.}

\item{thresh}{Integer, lower threshold for filtering CNVs.}
}
\value{
norm_data object
}
\description{
Short CNV stretches can be filtered out.
}

