% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{interpol}
\alias{interpol}
\title{Interpolate cluster means}
\usage{
interpol(theta, hcenters, vcenters)
}
\arguments{
\item{theta}{Vector of theta values.}

\item{hcenters}{Vector of length three with the horizontal cluster centers.}

\item{vcenters}{Vector of length three with the vertical cluster centers.}
}
\value{
Numeric of interpolated intensity value.
}
\description{
Interpolate cluster means
}
\keyword{internal}

