% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{require_package}
\alias{require_package}
\title{Require package wrapper}
\usage{
require_package(pkg)
}
\description{
Wrapper to require package and return an error, if it is missing.
}
\references{
http://r-pkgs.had.co.nz/description.html#dependencies
}
\keyword{internal}

