% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synteny.R
\name{split_sb}
\alias{split_sb}
\title{Split synteny blocks}
\usage{
split_sb(block, df, max, bnames, smoothing = TRUE)
}
\arguments{
\item{block}{Vector,}

\item{df}{Data.frame,}

\item{max}{Integer,}

\item{bnames}{Character,}

\item{smoothing}{Logical, if block positions should be smoothed}
}
\description{
Split synteny blocks
}
\keyword{internal}

