% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{syn_uni}
\alias{syn_uni}
\title{Brassica napus synteny data}
\format{A data.frame object, consisting of eight columns
\describe{
  \item{index1}{Chromosome index in first subgenome}
  \item{start1}{Position in first subgenome}
  \item{index2}{Chromosome index in second subgenome}
  \item{start2}{Position in second subgenome}
  \item{Alev}{Chromosome level}
  \item{Clev}{Chromosome level}
  \item{AGlo}{Global position}
  \item{CGlo}{Global position}
}}
\usage{
syn_uni
}
\description{
Synteny alignments created with DECIPHER using publicly available FASTA sequences
}
\keyword{datasets}

