% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synteny.R
\docType{class}
\name{synteny_info}
\alias{synteny_info}
\title{Class for synteny information}
\format{\code{\link{R6Class}} object.}
\usage{
synteny_info
}
\value{
Object of \code{\link{R6Class}}.
}
\description{
Class for synteny information
}
\keyword{data}

